Move the Fields Into Their Own Files
Since we're goign to add more and the original `fields.js` file was
200+ lines.
